+++ source/extensions/omni.isaac.orbit/test/sim/test_simulation_context.py
@@ -63,6 +63,7 @@ class TestSimulationContext(unittest.TestCase):
         # check valid settings
         self.assertEqual(sim.get_physics_dt(), cfg.dt)
         self.assertEqual(sim.get_rendering_dt(), cfg.dt * cfg.substeps)
+        self.assertFalse(sim.has_rtx_sensors())
         # check valid paths
         self.assertTrue(prim_utils.is_prim_path_valid("/Physics/PhysX"))
         self.assertTrue(prim_utils.is_prim_path_valid("/Physics/PhysX/defaultMaterial"))